+++ themes/ss_wap/tpl_search.php
@@ -25,7 +25,28 @@ list($seo_title,$seo_keywords,$seo_description) = ss_seo_render('search');
     <?php if ($recentread_url_raw !== ''): ?><a href="<?=$recentread_url_attr?>" rel="nofollow" class="bookcase">阅读记录</a><?php endif; ?>
     <h1>搜索小说</h1>
 </div>
-<div class="waps_r"><div class="search"><?php require_once 'tpl_search_form.php'; ?></div></div>
+<?php
+$search_url_raw = function_exists('ss_search_url') ? (string)ss_search_url() : '';
+$search_url_attr = ss_h($search_url_raw);
+?>
+<div class="waps_r"><div class="search">
+<form id="post" name="t_frmsearch" method="post" action="<?=$search_url_attr?>"<?php if ($search_url_raw === ''): ?> onsubmit="return false;"<?php endif; ?>>
+    <table cellpadding="0" cellspacing="0" style="width:100%;">
+        <tr>
+            <td style="width:50px;">
+                <div id="type" class="type">综合</div>
+            </td>
+            <td style="background-color:#fff; border:1px solid #CCC;">
+                <input id="s_key" name="searchkey" type="text" class="key" value="<?=htmlspecialchars($searchkey_safe, ENT_QUOTES, 'UTF-8')?>" placeholder="输入书名/作者" maxlength="50">
+                <input type="hidden" name="searchtype" value="all">
+            </td>
+            <td style="width:35px; background-color:#0080C0; background-image:url('/static/<?=$theme_dir_attr?>/search.png'); background-repeat:no-repeat; background-position:center">
+                <input name="t_btnsearch" type="submit" value="" class="go"<?php if ($search_url_raw === ''): ?> disabled="disabled" aria-disabled="true"<?php endif; ?>>
+            </td>
+        </tr>
+    </table><span id="s_tips"></span>
+</form>
+</div></div>
 <?php if($searchkey_safe !== ''): ?>
     <h1 class="search-title">搜索<span class="red"><?=htmlspecialchars($searchkey_safe, ENT_QUOTES, 'UTF-8')?></span>的结果<?php if (!empty($search_res) && is_array($search_res)): ?><small style="font-size:12px;color:#999;">（<?=count($search_res)?>条）</small><?php endif; ?></h1>
 <?php endif; ?>